[#36] feat: wishController 반환타입 제외 로직 작성

+++ backend/src/main/java/com/team14/cherrybnb/room/application/WishService.java
@@ -46,8 +46,11 @@ public class WishService {
     }
 
     // 위시 리스트에서 제거
-    public void removeWish(Long wishId) {
-        Wish wish = wishRepository.getById(wishId);
+    public void removeWish(Long wishId, Member loginMember) {
+        Wish wish = wishRepository.findById(wishId).orElseThrow(RuntimeException::new);
+        if (!loginMember.isSame(wish.getMember())) {
+            throw new RuntimeException();
+        }
         wishRepository.delete(wish);
     }
 }
